wristband_mqtt/requirements.txt
paho-mqtt
arrow==0.14.2
requests
configparser
Flask==1.0.2
cherrypy